types: move all types in own file types.py
closes #24
