+++ schulze_condorcet/__init__.py
@@ -1,4 +1,5 @@
 from .schulze_condorcet import (
-    pairwise_preference, schulze_evaluate, schulze_evaluate_detailed, Candidate, Vote)
+    pairwise_preference, schulze_evaluate, schulze_evaluate_detailed
+)
 
-__all__ = ['pairwise_preference', 'schulze_evaluate', 'schulze_evaluate_detailed', 'Candidate', 'Vote']
+__all__ = ['pairwise_preference', 'schulze_evaluate', 'schulze_evaluate_detailed']
