+++ schulze_condorcet/schulze_condorcet.py
@@ -1,34 +1,14 @@
 import itertools
 from gettext import gettext as _
 from typing import (
-    Collection, Container, Dict, List, Mapping, NewType, Tuple, TypedDict, Sequence
+    Collection, Container, List, Mapping, Tuple, Sequence
 )
 
-from schulze_condorcet.strength import StrengthCallback, winning_votes
-
-
-# A single vote, inputted as raw vote string. Consists only of the candidates,
-# separated by '>' and '=' to express the preference between the candidates.
-Vote = NewType('Vote', str)
-# A single candidate. Can be any string, expect that it must not contain the special
-# characters '>' and '='.
-Candidate = NewType('Candidate', str)
-# A single vote, split into separate levels accordingly to (descending) preference.
-# All candidates at the same level (in the same inner tuple) have equal preference.
-SplitVote = Sequence[Sequence[Candidate]]
-# How many voters prefer the first candidate over the second candidate.
-PairwisePreference = Dict[Tuple[Candidate, Candidate], int]
-# The link strength between two candidates.
-LinkStrength = Dict[Tuple[Candidate, Candidate], int]
-# The result of the schulze_condorcet method. This has the same structure as a SplitVote.
-SchulzeResult = List[List[Candidate]]
-
-
-class DetailedResultLevel(TypedDict):
-    preferred: List[Candidate]
-    rejected: List[Candidate]
-    support: PairwisePreference     # between pairs of preferred and rejected candidates
-    opposition: PairwisePreference  # between pairs of rejected and preferred candidates
+from schulze_condorcet.strength import winning_votes
+from schulze_condorcet.types import (
+    Candidate, DetailedResultLevel, LinkStrength, PairwisePreference, SchulzeResult,
+    SplitVote, StrengthCallback, Vote
+)
 
 
 def _schulze_winners(d: Mapping[Tuple[Candidate, Candidate], int],
